fix be and add frontend

+++ backend-springboot/src/main/java/com/nmquan1503/backend_springboot/dtos/responses/movie/MovieListItemResponse.java
@@ -19,7 +19,7 @@ public class MovieListItemResponse {
     LocalDate releasedDate;
     Short duration;
     Double averageRating;
-    Integer ratingCount;
+    Long ratingCount;
     List<CategoryResponse> categories;
     AgeRatingLabelResponse ageRating;
